+++ src/datasources/exchange-api/exchange-api.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { HttpErrorHandler } from '../errors/http-error-handler';
-import { ExchangeApi } from './exchange.service';
+import { ExchangeApi } from './exchange-api.service';
 
 @Module({
   providers: [HttpErrorHandler, ExchangeApi],
   exports: [ExchangeApi],
 })
-export class ExchangeModule {}
+export class ExchangeApiModule {}
